vengono create due stagioni e i relativi listini sia per la camera singola che per la camera doppia
git-svn-id: svn://93.63.174.84:8080/repos/locanda/trunk@640 5a08f537-7b09-473d-a1d1-551e4afdb7c2

+++ src/model/listini/ItemListinoCamera.java
@@ -3,7 +3,7 @@ package model.listini;
 public class ItemListinoCamera {
 	private Integer id;
 	private Integer numGuests;
-	private Integer[] prices = new Integer[7]; 
+	private Double[] prices = new Double[7]; 
 	//prezzo lun = prices[0]
 	//
 	//prezzo sabato = prices[5]
@@ -23,13 +23,12 @@ public class ItemListinoCamera {
 	public void setNumGuests(Integer numGuests) {
 		this.numGuests = numGuests;
 	}
-	public Integer[] getPrices() {
+	public Double[] getPrices() {
 		return prices;
 	}
-	public void setPrices(Integer[] prices) {
+	public void setPrices(Double[] prices) {
 		this.prices = prices;
 	}
-	
-	
+		
 
 }
